Remove outdated rc file for afsd on FBSD
The semantics of rc scripts have changed drastically since this
was written.  It will need to be rewritten from scratch; the
new version will live in src/packaging/FreeBSD as afsd.in, as it
is included with the port packaging in that form.

Reviewed-on: http://gerrit.openafs.org/4350
Reviewed-by: Derrick Brashear <shadow@dementia.org>
Tested-by: Derrick Brashear <shadow@dementia.org>
(cherry picked from commit dce071acef1119aceca7eb80a66c035fb03e9259)

Change-Id: I17bd69033e51bc5c59d70f982f8deb04fd18d5b4
Reviewed-on: http://gerrit.openafs.org/4354
Reviewed-by: Derrick Brashear <shadow@dementia.org>
Tested-by: Derrick Brashear <shadow@dementia.org>