OpenBSD: Fix parameters in call to afs_close()
OpenBSD passes an extra parameter in a call to afs_close()
in its OS-specific code. It went undetected in the past
but now causes compile headaches because of more accurate
prototypes in the header file. This patch removes the
offending extra parameter (which never actually caused any
problems because it was the last one in the parameter list).

Reviewed-on: http://gerrit.openafs.org/3685
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Derrick Brashear <shadow@dementia.org>
(cherry picked from commit 5ff6c88fcc0c35843e91dc45e4bd297efd0d4445)

Change-Id: I7d2df5fe80bd4de8c74200ada879a7bc1358e134
Reviewed-on: http://gerrit.openafs.org/3719
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Derrick Brashear <shadow@dementia.org>